Feat: Initial release of the new package
This package closely resembles `react-native-smooth-pincode-input` but
since that package is no longer maintained and isn't giving ownership
away, this serves as a replacement.

Its effectively a re-write of the original package but in Typescript and
with proper tests behind it.

It utilizes the exact same API as the original package and so is a drop
in replacement for it.